Port setup from `php-project`
