--- .codeclimate.yml
@@ -1,21 +1,0 @@
----
-engines:
-  duplication:
-    enabled: true
-    config:
-      languages:
-      - php
-  fixme:
-    enabled: true
-  phpmd:
-    enabled: true
-    config:
-      file_extensions:
-        - php
-      rulesets: "phpmd.xml"
-ratings:
-  paths:
-  - "**.inc"
-  - "**.php"
-exclude_paths:
-- test/
