+++ .env.example
@@ -1,0 +1,8 @@
+BASE_PATH=./
+BUILD_ROOT_PATH=/app/
+PROJECT_NAME=dhii_i18n-interface
+
+PHP_BUILD_VERSION=7.0
+PHP_TEST_VERSION=7.0
+
+
