+++ .gitattributes
@@ -1,9 +1,16 @@
-test export-ignore
+tests export-ignore
 nbproject export-ignore
+.idea export-ignore
 .codeclimate.yml export-ignore
 .php_cs export-ignore
+phpcs.xml export-ignore
+phpcs.xml.dist export-ignore
+psalm.xml export-ignore
+psalm.xml.dist export-ignore
 .travis.yml export-ignore
 .gitattributes export-ignore
 .gitignore export-ignore
 composer.lock export-ignore
 phpunit.xml export-ignore
+phpunit.xml.dist export-ignore
+docker export-ignore
