+++ .gitignore
@@ -1,5 +1,7 @@
 /vendor/
-/nbproject/private/
-/test/coverage/
-/test/log/
 /build/
+/.idea/workspace.xml
+/.idea/codeStyles/
+/.idea/inspectionProfiles/
+/.idea/misc.xml
+/.env
