--- .travis.yml
@@ -1,21 +1,0 @@
-language: php
-php:
-  - '5.3'
-  - '5.4'
-  - '5.5'
-  - '5.6'
-  - '7.0'
-  - '7.1'
-  - nightly
-
-before_script:
-  - composer update --prefer-dist
-
-script:
-  - vendor/bin/phpunit
-
-after_script:
-    - vendor/bin/test-reporter --coverage-report="test/coverage/clover.xml"
-cache:
-  directories:
-    - $HOME/.composer/cache/files
