+++ composer.json
@@ -11,19 +11,20 @@
     ],
     "minimum-stability": "dev",
     "require": {
-        "php": "^5.3 | ^7.0"
+        "php": "^7.0"
     },
     "require-dev": {
-        "dhii/php-cs-fixer-config": "dev-php-5.3",
-        "phpunit/phpunit": "^4.8",
-        "ptrofimov/xpmock": "^1.1",
-        "codeclimate/php-test-reporter": "<=0.3.2",
-        "dhii/stringable-interface": "^0.1",
-        "dhii/data-key-value-aware-interface": "^0.1"
+        "phpunit/phpunit": "^6.0 | ^7.0 | ^8.0 | ^9.0",
+        "ptrofimov/xpmock": "^1.1"
     },
     "autoload": {
         "psr-4": {
             "Dhii\\I18n\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Dhii\\I18n\\Test\\Func\\": "src/"
         }
     },
     "extra": {
