--- nbproject/project.properties
@@ -1,24 +1,0 @@
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_create_2e_tests=false
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.configuration_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.configuration_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.customSuite_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.customSuite_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.test_2e_groups_2e_ask=false
-auxiliary.org-netbeans-modules-php-phpunit.test_2e_run_2e_all=false
-file.reference.i18n-interface-vendor=vendor
-file.reference.test-functional=test/functional
-include.path=\
-    ${php.global.include.path}:\
-    ${file.reference.i18n-interface-vendor}
-php.version=PHP_53
-source.encoding=UTF-8
-src.dir=src
-tags.asp=false
-tags.short=false
-test.src.dir=${file.reference.test-functional}
-testing.providers=PhpUnit
-web.root=.
