--- src/Exception/I18nExceptionInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Dhii\I18n\Exception;
-
-/**
- * Something that can represent an exception which is related to internationalization.
- *
- * @since 0.1
- */
-interface I18nExceptionInterface
-{
-}
