+++ src/FormatTranslatorInterface.php
@@ -2,30 +2,21 @@
 
 namespace Dhii\I18n;
 
-use Dhii\Data\ValueAwareInterface as Value;
-
 /**
- * Something that can act as a translator, which can translate a string format.
- *
- * The {@see translate()} method throws a more specialized
- * {@see Dhii\I18n\Exception\FormatTranslationExceptionInterface}.
- *
- * @since 0.1
+ * Something that can act as a translator, which can translate a string format
  */
-interface FormatTranslatorInterface extends TranslatorInterface
+interface FormatTranslatorInterface
 {
     /**
-     * {@inheritdoc}
+     * Translates a string with context, interpolating parameters into the string.
      *
      * If values are given, MUST attempt to use them for interpolating placeholders in the string.
      *
-     * @since 0.1
-     *
-     * @param string|Stringable $format  The string format to translate
-     * @param array|null        $params  A map of values to use for placeholder substitution, if any.
-     * @param string|Value|null $context A context for the string, if any.
+     * @param string      $format  The string format to translate
+     * @param array|null  $params  A map of values to use for placeholder substitution, if any.
+     * @param string|null $context A context for the string, if any.
      *
      * @return string The translated string, with placeholders interpolated.
      */
-    public function translate($format, $params = null, $context = null);
+    public function translate(string $format, array $params = null, string $context = null);
 }
