--- test/functional/Exception/FormatTranslationExceptionInterfaceTest.php
@@ -1,53 +1,0 @@
-<?php
-
-namespace Dhii\I18n\Exception\FuncTest\Exception;
-
-use Xpmock\TestCase;
-
-/**
- * Tests {@see Dhii\I18n\Exception\FormatTranslationExceptionInterface}.
- *
- * @since 0.1
- */
-class FormatTranslationExceptionInterfaceTest extends TestCase
-{
-    /**
-     * The name of the test subject.
-     *
-     * @since 0.1
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\\I18n\\Exception\\FormatTranslationExceptionInterface';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since 0.1
-     *
-     * @return Dhii\I18n\Exception\FormatTranslationExceptionInterface
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getSubject()
-            ->getTranslator()
-            ->getContext()
-            ->getParams()
-            ->new();
-
-        return $mock;
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since 0.1
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the test subject could not be created');
-        $this->assertInstanceOf('Dhii\\I18n\\Exception\\I18nExceptionInterface', $subject, 'Subject does not implement required interface');
-        $this->assertInstanceOf('Dhii\\I18n\\Exception\\TranslationExceptionInterface', $subject, 'Subject does not implement required interface');
-    }
-}
